Refactor global leaderboard into own service
